requirements.txt
streamlit>=1.56.0
ultralytics>=8.4.4
opencv-python-headless>=4.10.0
pillow>=10.0.0
pandas>=2.2.0
numpy>=1.26.0
``